validation during user input #2
