+++ app/src/main/java/org/dieschnittstelle/mobile/android/skeleton/MainActivity.java
@@ -78,68 +78,9 @@ public class MainActivity extends AppCompatActivity {
                     progressBar = findViewById(R.id.progressBar);
                     progressBar.setVisibility(View.GONE);
 
-                    emailAdressValidation();
-                    passwordValidation();
-
                     break;
             }
         }
-    }
-
-    private void emailAdressValidation() {
-        emailAdress.addTextChangedListener(new TextWatcher() {
-            private Handler emailAdressValidationHandler = new Handler(Looper.getMainLooper());
-
-            @Override
-            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-            @Override
-            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                viewModel.getMailInputError().setValue(null);
-                emailAdressValidationHandler.removeCallbacksAndMessages(null); // Remove pending validations
-                emailAdressValidationHandler.postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (viewModel.checkMailInput()) {
-                            viewModel.getMailInputError().setValue(null); // Clear any previous error
-                        } else {
-                            viewModel.getMailInputError().setValue("Input not valid, must be an e-mail address");
-                        }
-                    }
-                }, 2000);
-            }
-        });
-    }
-
-    private void passwordValidation() {
-        password.addTextChangedListener(new TextWatcher() {
-            private Handler passwordAdressValidationHandler = new Handler(Looper.getMainLooper());
-
-            @Override
-            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-            @Override
-            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {}
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                viewModel.getPasswordInputError().setValue(null);
-                passwordAdressValidationHandler.removeCallbacksAndMessages(null); // Remove pending validations
-                passwordAdressValidationHandler.postDelayed(new Runnable() {
-                    @Override
-                    public void run() {
-                        if (viewModel.checkPasswordInput()) {
-                            viewModel.getPasswordInputError().setValue(null); // Clear any previous error
-                        } else {
-                            viewModel.getPasswordInputError().setValue("Input not valid, must consist of 6 numbers");
-                        }
-                    }
-                }, 2000);
-            }
-        });
     }
 
     private void handleUserLoginState(MainViewModel.LoginState loginState) {
